requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.10.2
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
configobj==5.0.8
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
databricks-cli==0.17.5
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
docker==6.0.1
dpath==2.1.4
dulwich==0.21.3
dvc==2.48.0
dvc-data==0.43.0
dvc-http==2.30.2
dvc-objects==0.21.1
dvc-render==0.2.0
dvc-studio-client==0.6.0
dvc-task==0.2.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.1
Flask==2.2.3
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.39.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
funcy==1.18
gitdb==4.0.10
GitPython==3.1.31
grandalf==0.8
gunicorn==20.1.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
iterative-telemetry==0.0.7
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
kiwisolver==1.4.4
kombu==5.2.4
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mlflow==2.2.2
multidict==6.0.4
nanotime==0.5.2
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.8.7
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
polars==0.15.16
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pydot==1.4.2
pygit2==1.11.1
Pygments==2.13.0
pygtrie==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
querystring-parser==1.2.4
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
scmrepo==0.1.15
Send2Trash==1.8.0
shap==0.41.0
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.6
sqlparse==0.4.3
sqltrie==0.3.0
stack-data==0.6.2
tabulate==0.9.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomlkit==0.11.6
torch==1.13.1
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.15
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.3
widgetsnbextension==4.0.4
yarl==1.8.2
zc.lockfile==3.0.post1
zipp==3.15.0
